chore(baseline-refresh): thin story-close into a phase pipeline (resolves #2489)
