+++ tests/scripts/fixtures/story-close-pipeline/exports.json
@@ -1,0 +1,13 @@
+{
+  "exports": [
+    "captureStartingBranch",
+    "checkCdOutGuard",
+    "describeAutoRefreshOutcome",
+    "renderCoverageTimeoutFrictionBody",
+    "renderSpawnTimeoutFrictionBody",
+    "resolveSpawnTimeoutReason",
+    "restoreStartingBranch",
+    "runStoryClose",
+    "runStoryClosePreflight"
+  ]
+}
