+++ tests/scripts/fixtures/story-close-pipeline/no-args-stderr.txt
@@ -1,0 +1,1 @@
+[Orchestrator] ❌ [phase=fatal] [story-close] Error: Usage: node story-close.js --story <STORY_ID> [--epic <EPIC_ID>]
